Add support for reading data through HCatalog.
Reads through HCatalog use the existing Hive support, but need some changes. Specifically, metadata stored in the job conf is different between the two systems, so HiveMultiInputFormat needs to handle both styles of getting the serialization class name from the conf.

Full debug builds are enabled so a debugger works correctly.

A ReporterWrapper is added because HCatalog implements its own reporter, instead of extending StatusReporter. TODO: Figure out if this wrapper should stay, or update HCatalog to extend StatusReporter.

As an HCatalog dependency was added to read the serialization class, ivy has been updated. HCatalog is not yet in a maven repo so the jar is checked in. Hoping to remove this jar before merging to master.

Related to HCatalog's ProgressReporter, counters do not appear to work correctly. A workaround has been added but counters should in fact work.